feat: Implement rotation functionality for master key and secrets management
